refactoring state iwth usecontext. refactoring prop drilling in ingredientinputdisabled and faq
