+++ src/components/App/AppStyles.js
@@ -1,0 +1,16 @@
+const appSx = {
+  backgroundColor: 'background.default',
+  backgroundRepeat: 'repeat',
+  display: 'flex',
+  height: '100%',
+  justifyContent: 'center',
+  width: '100%',
+};
+
+const mainSx = {
+  display: 'flex',
+  justifyContent: 'center',
+  padding: { xs: '32px 0px', sm: '32px 16px' },
+};
+
+export { appSx, mainSx };
