+++ src/components/RecipesContext/RecipesContext.js
@@ -1,0 +1,22 @@
+// import React, { createContext } from 'react';
+// import PropTypes from 'prop-types';
+
+// export const RecipesContext = createContext({});
+
+// export const RecipesContextProvider = (props) => {
+//   const { recipesContextValue, children } = props;
+
+//   return (
+//     <RecipesContext.Provider value={{ recipesContextValue }}>
+//       {children}
+//     </RecipesContext.Provider>
+//   );
+// };
+
+// RecipesContextProvider.propTypes = {
+//   recipesContextValue: PropTypes.shape({
+//     state: PropTypes.object,
+//     dispatch: PropTypes.func,
+//   }),
+//   children: PropTypes.element.isRequired,
+// };
